Move compute function and rename file to merge_and_fill_utils.py
- Moved compute function to pvnet_uk_regional.py
- Renamed process_and_combine.py to merge_and_fill_utils.py
- Updated all related imports
