+++ ocf_data_sampler/torch_datasets/pvnet_uk_regional.py
@@ -18,10 +18,9 @@ from ocf_data_sampler.numpy_sample import (
     convert_gsp_to_numpy_sample,
     make_sun_position_numpy_sample,
 )
-from ocf_data_sampler.torch_datasets.process_and_combine import (
+from ocf_data_sampler.torch_datasets.merge_and_fill_utils import (
     merge_dicts,
     fill_nans_in_arrays,
-    compute,
 )
 from ocf_data_sampler.numpy_sample.gsp import GSPSampleKey
 from ocf_data_sampler.numpy_sample.nwp import NWPSampleKey
