+++ tests/torch_datasets/test_process_and_combine.py
@@ -1,6 +1,6 @@
 import numpy as np
 
-from ocf_data_sampler.torch_datasets.process_and_combine import (
+from ocf_data_sampler.torch_datasets.merge_and_fill_utils import (
     merge_dicts,
     fill_nans_in_arrays,
 )
